moved TEST_USER to demo user generator
