+++ src/app/core/import/import/import-sample-raw-data.ts
@@ -5,8 +5,7 @@ import { Entity } from "../../entity/model/entity";
 import { ColumnMapping } from "../column-mapping";
 import { genders } from "../../../child-dev-project/children/model/genders";
 import { ImportMetadata } from "../import-metadata";
-
-import { TEST_USER } from "../../../utils/mock-local-session";
+import { TEST_USER } from "../../user/demo-user-generator.service";
 
 /**
  * Sample raw data that can be used in Storybook and tests.
