+++ src/app/core/session/auth/local/local-auth.service.spec.ts
@@ -1,6 +1,6 @@
 import { LocalAuthService } from "./local-auth.service";
 import { SessionInfo } from "../session-info";
-import { TEST_USER } from "../../../../utils/mock-local-session";
+import { TEST_USER } from "../../../user/demo-user-generator.service";
 
 describe("LocalAuthService", () => {
   let service: LocalAuthService;
