+++ src/app/core/support/support/support.component.spec.ts
@@ -18,13 +18,13 @@ import { NoopAnimationsModule } from "@angular/platform-browser/animations";
 import { PouchDatabase } from "../../database/pouch-database";
 import { BackupService } from "../../admin/backup/backup.service";
 import { DownloadService } from "../../export/download-service/download.service";
-import { TEST_USER } from "../../../utils/mock-local-session";
 import { SyncService } from "../../database/sync.service";
 import { KeycloakAuthService } from "../../session/auth/keycloak/keycloak-auth.service";
 import { SyncStateSubject } from "../../session/session-type";
 import { User } from "../../user/user";
 import { CurrentUserSubject } from "../../session/current-user-subject";
 import { SessionSubject } from "../../session/auth/session-info";
+import { TEST_USER } from "../../user/demo-user-generator.service";
 
 class MockDeleteRequest {
   onsuccess: () => {};
