+++ src/app/core/user/demo-user-generator.service.ts
@@ -3,13 +3,15 @@ import { Injectable } from "@angular/core";
 import { User } from "./user";
 import { faker } from "../demo-data/faker";
 
+export const TEST_USER = "test";
+
 /**
  * Generate demo users for the application with its DemoDataModule.
  */
 @Injectable()
 export class DemoUserGeneratorService extends DemoDataGenerator<User> {
   /** the username of the basic account generated by this demo service */
-  static DEFAULT_USERNAME = "demo";
+  static DEFAULT_USERNAME = TEST_USER;
   static ADMIN_USERNAME = "demo-admin";
 
   /**
