--- src/app/utils/mock-local-session.ts
@@ -1,2 +1,0 @@
-export const TEST_USER = "test";
-export const TEST_PASSWORD = "pass";
